Refactor to avoid for loops on strings and hide Kaiko concerns in KaikoDataReader
